refactor: update DeleteTaskUseCase to use direct calls & the related imports
