+++ src/main/kotlin/di/UseCaseModule.kt
@@ -3,7 +3,7 @@ package di
 import logic.useCase.CreateTaskUseCase
 import org.example.logic.useCase.*
 import org.example.logic.useCase.DeleteProjectUseCase
-import org.example.logic.useCase.deleteTask.DeleteTaskUseCase
+import org.example.logic.useCase.DeleteTaskUseCase
 import org.example.logic.useCase.UpdateProjectUseCase
 import org.example.logic.useCase.UpdateTaskUseCase
 import org.koin.core.module.dsl.singleOf
