+++ src/main/kotlin/logic/useCase/DeleteTaskUseCase.kt
@@ -1,16 +1,11 @@
-package org.example.logic.useCase.deleteTask
+package org.example.logic.useCase
 
 import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
-import org.example.logic.command.CreateAuditLogCommand
-import org.example.logic.command.TransactionalCommand
 import org.example.logic.models.AuditLog
 import org.example.logic.models.AuditLogActionType
 import org.example.logic.models.AuditLogEntityType
 import org.example.logic.repositries.AuditLogRepository
 import org.example.logic.repositries.TaskRepository
-import org.example.logic.useCase.GetCurrentUserUseCase
-import org.example.logic.useCase.GetTaskByIdUseCase
 import org.example.logic.utils.UnableToDeleteTaskException
 import org.example.logic.utils.formattedString
 import org.example.logic.utils.getCroppedId
@@ -21,13 +16,15 @@ class DeleteTaskUseCase(
     private val taskRepository: TaskRepository,
     private val auditLogRepository: AuditLogRepository,
     private val userUseCase: GetCurrentUserUseCase,
-    private val getTaskByIdUseCase: GetTaskByIdUseCase,
 ) {
     suspend operator fun invoke(taskId: String) {
-        val auditLog = saveAuditLog(taskId)
-        val auditCommand = CreateAuditLogCommand(auditLogRepository, auditLog)
-        val deleteTasksCommand = DeleteTaskCommand(taskRepository, getTaskByIdUseCase(taskId))
-        TransactionalCommand(listOf(auditCommand, deleteTasksCommand), UnableToDeleteTaskException("Cannot delete task"))
+        try {
+            auditLogRepository.createAuditLog(saveAuditLog(taskId))
+
+            taskRepository.deleteTask(taskId)
+        } catch (e: Exception) {
+            throw UnableToDeleteTaskException("Cannot delete task")
+        }
     }
 
     @OptIn(ExperimentalUuidApi::class)
