+++ src/test/kotlin/presentation/screens/ShowTaskInformationTest.kt
@@ -8,7 +8,7 @@ import org.example.logic.models.Task
 import org.example.logic.useCase.GetEntityAuditLogsUseCase
 import org.example.logic.useCase.GetStateNameUseCase
 import org.example.logic.useCase.GetTaskByIdUseCase
-import org.example.logic.useCase.deleteTask.DeleteTaskUseCase
+import org.example.logic.useCase.DeleteTaskUseCase
 import org.example.logic.useCase.UpdateTaskUseCase
 import org.example.presentation.screens.ShowTaskInformation
 import org.junit.jupiter.api.BeforeEach
